+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Symbol information is now provided to the Sonar API for use in the web interface.
 - `MemoryManagementRule` analysis rule, which flags object allocations without memory management.
 - `ShortIdentifiersRule` analysis rule, which flags identifiers below a certain length.
+- `ForbiddenIdentifierRule` analysis rule template, which flags identifiers with a specified name.
 
 ### Changed